This is an interim commit; not ready for merging; but to support code sharing.
This is an in-progress effort to create a monolithic version of the POP brain.  We are not ready aiming to give up all the goodness of the distributed AWS technology, but this does provide a potentially convenient means of demo'ing and prototyping on box.

Again, this is a work in-progress; not ready to be demo'ed just yet.