This is an interim commit; not ready for merging; but to support code sharing.
This is an in-progress effort to create a monolithic version of the POP brain.  We are not ready aiming to give up all the goodness of the distributed AWS technology, but this does provide a potentially convenient means of demo'ing and prototyping on box.

Again, this is a work in-progress; not ready to be demo'ed just yet.

However, to call out that we want a functional monolithic brain, we have an integration test MonolithicApplicationTest.runAgenda.  This test exercises much of the brain activity, including: creation and persistence of endpoints; checking for authority to use compute resources (e.g. resource pools - typically would represent a Kubernetes cluster) as well as capability of compute resources to execute the requested operations; this also includes populating agendaready queues.

Agenda progress logic (via the monolith) is still not confirmed.

Additionally, we have not yet integrated the monolith brain with an actual resource pool.

These two preceeding steps need to be done before we can really get into producing tutorials.